requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
yarl==1.9.3
